update new version

+++ app/Http/Controllers/CategoriesController.php
@@ -54,6 +54,7 @@ class CategoriesController extends Controller
     //store category
     public function store(Request $request)
     {
+        $this->authorize('create-update-categories');
         $new = new Categories();
         try {
             $validatedData = Validator::make($request->all(), [
@@ -86,7 +87,7 @@ class CategoriesController extends Controller
     //update
     public function update(Request $request, $id)
     {
-
+        $this->authorize('create-update-categories');
         $category = Categories::where('id', $id)->first();
         if ($category == "") {
             return response()->json([
@@ -110,6 +111,7 @@ class CategoriesController extends Controller
 
     public function delete($id)
     {
+        $this->authorize('create-update-categories');
         $category = Categories::findOrFail($id);
         $category->delete();
         return response()->json([
